meal prediction - backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0

